nas-algorithm/requirements.txt
certifi==2022.12.7
charset-normalizer==3.3.2
common==0.1.2
facenet-pytorch==2.5.3
idna==3.4
joblib==1.3.2
nats-py==2.6.0
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.4.0.42
pandas==1.3.5
Pillow==9.5.0
pip==22.3.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==0.22.2.post1
scipy==1.7.3
setuptools==65.6.3
six==1.16.0
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.7.1
urllib3==2.0.7
wheel==0.38.4
